requirements.txt
# Core dependencies
playwright>=1.40.0
litellm>=1.30.0
requests>=2.31.0
PyYAML>=6.0.1

# LLM response parsing
json-repair>=0.25.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Retry logic
tenacity>=8.2.0

# Geo/mapping utilities
polyline>=2.0.0
shapely>=2.0.0

# Database
lmdb>=1.4.0

# Optional: Azure identity (for TRAPI proxy)
# azure-identity>=1.15.0

